fix hungarian matcher

+++ src/otx/algo/common/utils/assigners/hungarian_matcher.py
@@ -279,7 +279,7 @@ class HungarianMatcher(nn.Module):
                 "pred_boxes": outputs["pred_boxes"][i],
                 "pred_masks": outputs["pred_masks"][i] if "pred_masks" in outputs else None,
                 "target_boxes": targets[i]["boxes"],
-                "target_labels": targets[i]["labels"],
+                "target_labels": targets[i]["labels"].long(),
                 "target_mask": targets[i]["masks"] if "masks" in targets[i] else None,
             }
             for i in range(batch_size)
